refactor: split wiki attachment list into list.rs per project convention
Addresses review comment: mod.rs should be module wiring only, not implement commands
